+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add new Facade class `Apiato` containing the old Butlers classes functions, in addition to the `call` magical methode (`Apiato::call()`) in the `CallableTrait`.
 - Add container specific config file to each container.
 - Add readme.md file to each container.
+- Add possibility to `customize` the way, `Exceptions` are displayed (cf. `ExceptionFormatters`)
 
 ### Changed
 - Upgrade Mail to use Laravel 5.5 Mails. And add Ship directory for Mail in addition to user mail sample in the User container.
@@ -48,6 +49,7 @@
 - Renamed `wepay.php` config file to `wepay-container.php`.
 - Rename `ListAll` to `GetAll` in every Actions/Tasks/controller functions/route files/requests
 - Rename `Get` to `Find` in every Actions/Tasks/controller functions/route files/requests
+- Slight adaptations to the `Exception` message format (due to `ExceptionFormatters`)
 
 ### Fixed
 - Fix Social Authentication errors.